+++ CHANGELOG.md
@@ -12,6 +12,7 @@ All notable changes to this project will be documented in this file.
 - Added field `ttl` to the `cache` processor and output.
 - New `sql` output, which is similar to the `sql` processor and currently supports Clickhouse, PostgreSQL and MySQL.
 - The `kafka` input now supports multiple topics, topic partition balancing, and checkpointing.
+- New `cassandra` output.
 
 ## Changed